- Code by G.Artim: Hello,
I was recently using your add on. I had to implement a few new features and thought you might be interested in using them as I noticed some people were requesting some new features. I implemented an additional timer to allow for any AJAX/javascript onload functions to run. I added a default option as well as the ability to choose the time you wish to wait manually during the prompt. I also noticed that in your reports you were printing out the result termniated by a semi-colon, I changed that over to a comma so that each value had its own cell in excel when it was opened. Also I was having an issue that when I set the max page load time to 0 it wouldn't run at all. I know it is supposed to remove any time out, but instead it was just closing the application. To fix that issue I just added 1.5 seconds to the timeout so that atleast the program ran. If you have any questions please feel free to contact me and I will respond as soon as I can. Thank you.

Sincerely,
Gregory Artim

- Code by Rafal Zelazko
 - redesign
 - minor bugfixes
 - update install.rdf